Don't recycle WAL when receiving archival report in archive_mode always (#1183)
Commit 4f2db19 added the following functionality. If
wal_sender_archiving_status_interval GUC was set, the primary would
periodically (by default, every 10 seconds) notify its standby about the latest
WAL segment sent to the archive. At the same time, the standby would mark all
old WAL segments as archived (.done), regardless of the archiving mode, upon
receiving such a notification. As a result, the WAL segment could be reused and
overwritten (or even deleted!), in parallel with archiving from the standby,
which could lead to a broken WAL segment getting into the archive. A similar
problem was fixed in commit 4b818bd, but not for the notification case, so this
patch solves this problem exactly for this case. Namely, upon receiving a
notification from the primary, it does not mark WAL segments as archived
(.done) in the always archiving mode.

Ticket: ADBDEV-6950