Fix creating table by EXPLAIN ANALYZE CTAS from prepared statement (#1214)
When creating a table from a select statement, GPDB uses the following hack.
The intorel_initplan function is called in InitPlan before ExecutorRun in
ExecCreateTableAs so that the table is created earlier, otherwise the current
MPP framework will fail. This only happens if the queryDesc->plannedstmt
structure has the intoClause field filled in. For EXPLAIN ANALYZE CTAS from a
prepared statement, this field was forgotten to be filled in, the
intorel_initplan function was not called, the table was not created, and the
custom DestReceiver DR_intorel was not initialized correctly. This patch solves
these problems by correctly setting the intoClause field of the plannedstmt
structure in the ExplainOnePlan function, before calling the ExecutorStart
function, which later calls InitPlan, which then calls intorel_initplan.

Ticket: ADBDEV-7030
(cherry picked from commit 7b908315de85f1671b6733fbe6aadb4f2cfcface)

Changes from original commit:
1) rework solution to provide more correct assignment with copy object
2) removed unnecessary changes in intorel_startup_dummy function
3) matchsubs were added to tests, becasuse 6X doesn't support SUMMARY OFF
   option in EXPLAIN
4) missed output for ORCA was added, because now it includes plan