Fix Orca's planning of a correlated subplan that has distinct or union all (#1246)
For correlated subqueries in any/in/exists expressions, ORCA tries to build an
uncorrelated plan, but if it fails, it builds a correlated one. The
XformSelect2Apply transformation creates two versions of the expression with
CLogicalLeftSemiApplyIn and CLogicalLeftSemiCorrelatedApplyIn. For both
expressions, the XformLeftSemiApply2LeftSemiJoin transformation is applied, in
which an attempt is made to decorrelate the subquery, which, if failed, leads to
the decoration of the subquery by the Result node with the true constant. For
CLogicalLeftSemiApplyIn, this happens in the
PexprLogicalCorrelatedQuantifiedApply method. If there is no filter at the top
of the expression, then the Result node with the constant true is added. In this
place, there was an assert that checked the presence of external parameters. But
the presence of external parameters in this place is to be expected.

Fix this problem by removing the assert, as the presence of external parameters
in this location is expected. Also add a new GUC that controls the
XformSelect2Apply transformation in order not to generate a correlated
CLogicalLeftSemiCorrelatedApplyIn expression; this is necessary for tests.