Adapt changes from 7.x to allow release resources from UDF SRF. (#1222)
Adapt changes from 7.x to allow release resources from UDF SRF (#1222).
This commit adapts changes from commit b5df840 (#771) to 7.x.

Previously, in 7.x, this was handled by executing a Squelch command on the
ProjectSet node. However, in 6.x, there is no ProjectSet, and function calls are
executed differently.

In this implementation, a chain of resource release callbacks is used to invoke
the UDF one last time in Squelch mode.

Problem:
some Greenplum functions, f.e. __gp_aoblkdir and __gp_aovisimap do not release
resources if used in context with limit clause.

Cause:
It is possible for Set Returning Function (SRF) to work in ValuePerCall
mode. But according to the PostgreSQL documentation the query is not guaranteed
to be run to completion; that is, due to options such as LIMIT, the executor
might stop making calls to the set-returning function before all rows have been
fetched. This means it is not safe to perform cleanup activities in the last
call, because that might not ever happen. It's recommended to use Materialize
mode for functions that need access to external resources, such as file
descriptors.

However when running in Materialize mode SRF can consume a lot of resources
which will be not needed in case with LIMIT. F.e. __gp_aoblkdir is expected to
return 1 row for minipage. It means ~130K rows per 1TB of storage. This can lead
to overhead in ~7Mb for 1Tb of disk storage. This memory will be allocated on
segments and not used. To address this issue Greenplum introduced so called
Squelch function to States. Unfortunately Squelch is not supported for the
SRF's.

Fix: This patch adds support of squelching for SRFs with extension of
ValuePerCall protocol. To determine if the function called to release resources
this patch introduces macro SRF_IS_SQUELCH_CALL(). This macro calls function
is_SquelchFuncCall and shall be checked at the very beginning of SRF. It
performs 2 things:

    it sets flag SFRM_Squelch in the returnMode fields to mark
    this function as supporting squelch at all
    it checks for flag SFRM_Squelch allowedModes - in this case function is
    called in squelch mode to free resources.

This implementation minimized changes to ABI, and also allows functions which
don't need squelching to be not touched.