Fix permission error for synchronization SET (#1397)
993b6c4 changed how SIGHUP synchronization works by replacing SET commands
with a call to pg_catalog.set_config(). It was left unnoticed that the latter
method does not bypass permissions, as it should, based on the distributed
transaction flags, to allow changing all GUCs with context equal to or lower
than PGC_SIGHUP. The bypass was left implemented only for SET and there was no
tests to cover such behavior.

Since SET isn't used for sync anymore, the same check is moved to
pg_catalog.set_config(). A test is added to make sure the gang and a temporary
table remains intact after sending the signal multiple times, indicating that no
error has occurred.

Ticket: ADBDEV-7318