Eliminate SSL renegotiation from gpfdist (#1237)
SSL renegotiation is considered insecure and vulnerable and was eliminated in
TLS v1.3.

However, when using older clients, we want to ensure that renegotiation is not
used, so we explicitly disable it whenever possible.

This commit removes SSL renegotiation setup from gpfdist.

Additionally, disabling SSL renegotiation in this commit resolves an issue with
TLS v1.2, specifically an incompatibility between OpenSSL 1.0.2k and GnuTLS
during server-initiated renegotiation.

Testing this incompatibility requires a non-trivial setup, involving gpfdist
built with OpenSSL 1.0.2k and a client (ADB or curl) built with GnuTLS. Since
this setup is not achievable using the current build and test system, only
manual setup and testing are possible.