Fix errors on planning queries with zero column AO tables (#1567)
Problem description:
1. Planning a query with an AO column-oriented table with 0 columns caused an
error "floating-point exception", if Postgres planner was used.
2. Planning a query with an AO row-oriented table with 0 columns with GUC
'gp_enable_relsize_collection' enabled caused a failed assertion in function
'GetSegFilesTotalsCluster()'.

Both issues are a side effect of the commit
1f4d9b473693d91c4219a239456aa60ead0f5f7c.

Root cause:
1. The "floating-point exception" was caused by a division by zero when
calculating row count, because data width is zero when there are no columns in
a relation. The issue didn't occur for heap or row-oriented tables because data
width for them is adjusted by additional header sizes.
2. The assertions in 'GetSegFilesTotalsCluster()' were wrong, as actually the
SPI result of the query, used in the function, can return NULL values if there 
are no columns in a relation.

Fix:
1. Do the calculation for the tuple count only if the data width is not 0;
otherwise just set the tuple count to 1.
2. Remove assertions in 'GetSegFilesTotalsCluster()' where actually NULLs can
be returned.